refactor: 💡 fixed the joiners auto counter
fixed the joiners auto counter
