+++ src/pages/index.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import type { NextPage } from 'next';
-// import Mailchimp from 'mailchimp-api-v3';
+import type { GetStaticProps, NextPage } from 'next';
+import Mailchimp from 'mailchimp-api-v3';
 
 import Main from '@/containers/Main';
 
-// interface IListMembersData {
-// 	readonly stats: { readonly member_count: number };
-// }
+interface IListMembersData {
+	readonly stats: { readonly member_count: number };
+}
 
 interface IProps {
 	readonly joinersCount: number;
@@ -19,21 +19,43 @@ const Home: NextPage<IProps> = (props) => {
 Home.displayName = 'Home';
 Home.defaultProps = {};
 
-// export const getStaticProps: GetStaticProps = async () => {
-// 	const mailchimp = new Mailchimp(process.env.MAILCHIMP_API_KEY);
+export const getStaticProps: GetStaticProps = async () => {
+	const mailchimp = new Mailchimp(process.env.MAILCHIMP_API_KEY);
 
-// 	try {
-// 		const listMembersData: IListMembersData = await mailchimp.get(
-// 			`/lists/${process.env.MAILCHIMP_LIST_ID}`,
-// 		);
+	const startedDayDate = new Date('2022-03-14').getDate();
+	const startedMonthDate = new Date('2022-03-14').getMonth() + 1;
+	const currentDayDate = new Date().getDate();
+	const currentMonthDate = new Date().getMonth() + 1;
 
-// 		return {
-// 			props: { joinersCount: listMembersData.stats.member_count },
-// 			revalidate: 25200,
-// 		};
-// 	} catch {
-// 		return console.log('error');
-// 	}
-// };
+	// Coutner increased every day in the number of days that have been passed since the beginning of the count multiplied 17.
+	//Counter set up every day anew.
+	let counter = ((currentMonthDate - startedMonthDate) * 30 + (currentDayDate - startedDayDate)) * 17;
+
+	const updateCounter = () => {
+		counter += 17;
+
+		return counter;
+	};
+
+	updateCounter();
+
+	try {
+		const listMembersData: IListMembersData = await mailchimp.get(
+			`/lists/${process.env.MAILCHIMP_LIST_ID}`,
+		);
+
+		return {
+			props: { joinersCount: listMembersData.stats.member_count + counter },
+			//revalidate every 24 hours
+			revalidate: 86400,
+		};
+	} catch {
+		return {
+			props: { joinersCount: counter },
+			//revalidate every 24 hours
+			revalidate: 86400,
+		};
+	}
+};
 
 export default Home;
